requirements.txt
asgiref==3.5.2
cffi==1.15.1
cryptography==38.0.4
distlib==0.3.6
Django==4.1.1
django-cors-headers==3.13.0
django-rest-knox==4.2.0
djangorestframework==3.14.0
filelock==3.8.0
Pillow==9.3.0
platformdirs==2.5.2
psycopg2==2.9.5
pycparser==2.21
pytz==2022.2.1
sqlparse==0.4.2
tzdata==2022.2
virtualenv==20.16.5
virtualenvwrapper-win==1.2.7
